Code cleanup based on phpmd static analysis
- Enables additional phpmd rules and makes existing stricter
- Light refactor to Schema and SchemaProperty to reduce SA scores
