+++ src/Lib/Swagger.php
@@ -22,7 +22,6 @@ use Symfony\Component\Yaml\Yaml;
  * Class Swagger
  *
  * @package SwaggerBake\Lib
- * @SuppressWarnings(PHPMD.ExcessiveClassComplexity)
  */
 class Swagger
 {
